Update GUIslice_drv_tft_espi.cpp to work with Touch Panel CTS820_MUTUAL find on Lilygo T Display S3
This board Lilygo T Display S3 is in 2 variants. One with touchscreen and one without it.
The one with touch has little documentation on the net related to the touch panel, so I used the TouchLib library that can be found on the github page of the Lilygo product to create a driver. tested and works